chore(exporters/elasticsearch-exporter): handle missing ES cluster on open
- moves creating index templates to export phase
- adds fault tolerance test
